order-service/requirements.txt
fastapi
uvicorn
requests
sqlalchemy
pydantic
psycopg2-binary
python-dotenv